pyproject.toml
[project]
name = "mercury-mcp"
version = "0.1.0"
description = "MCP server for Mercury banking API"
requires-python = ">=3.8"
dependencies = [
    "dedalus-mcp",
    "pydantic",
    "requests",
    "python-dotenv",
]

[project.scripts]
mercury-mcp = "src.main:run"
